Moving operation overloading and testss to impl_ops module.
Adding documentation to impl macros

Moving tests into impl_ops module
