+++ rusty-machine/src/linalg/matrix/mod.rs
@@ -4,7 +4,7 @@
 //! relating to the matrix linear algebra struct.
 
 use std::fmt;
-use std::ops::{Mul, Add, Div, Sub, Index, IndexMut, Neg};
+use std::ops::{Mul, Add, Div, Sub, Neg};
 use libnum::{One, Zero, Float, FromPrimitive};
 use std::cmp::{PartialEq, min};
 use linalg::Metric;
@@ -12,6 +12,7 @@ use linalg::vector::Vector;
 use linalg::utils;
 
 mod decomposition;
+mod impl_ops;
 pub mod slice;
 
 /// Matrix dimensions
@@ -31,6 +32,34 @@ pub struct Matrix<T> {
     rows: usize,
     cols: usize,
     data: Vec<T>,
+}
+
+/// A MatrixSlice
+///
+/// This struct provides a slice into a matrix.
+///
+/// The struct contains the upper left point of the slice
+/// and the width and height of the slice.
+#[derive(Debug, Clone, Copy)]
+pub struct MatrixSlice<T> {
+    ptr: *const T,
+    rows: usize,
+    cols: usize,
+    row_stride: usize,
+}
+
+/// A mutable MatrixSlice
+///
+/// This struct provides a mutable slice into a matrix.
+///
+/// The struct contains the upper left point of the slice
+/// and the width and height of the slice.
+#[derive(Debug)]
+pub struct MatrixSliceMut<T> {
+    ptr: *mut T,
+    rows: usize,
+    cols: usize,
+    row_stride: usize,
 }
 
 impl<T> Matrix<T> {
@@ -109,22 +138,22 @@ impl<T> Matrix<T> {
     /// let a = Matrix::new(3,3, vec![2.0; 9]);
     /// let (b,c) = a.split_at(1, Axes::Row);
     /// ```
-    pub fn split_at(&self, mid: usize, axis: Axes) -> (slice::MatrixSlice<T>, slice::MatrixSlice<T>) {
-        let slice_1 : slice::MatrixSlice<T>;
-        let slice_2 : slice::MatrixSlice<T>;
+    pub fn split_at(&self, mid: usize, axis: Axes) -> (MatrixSlice<T>, MatrixSlice<T>) {
+        let slice_1 : MatrixSlice<T>;
+        let slice_2 : MatrixSlice<T>;
 
         match axis {
             Axes::Row => {
                 assert!(mid < self.rows);
 
-                slice_1 = slice::MatrixSlice::from_matrix(self, [0,0], mid, self.cols);
-                slice_2 = slice::MatrixSlice::from_matrix(self, [mid,0], self.rows - mid, self.cols);
+                slice_1 = MatrixSlice::from_matrix(self, [0,0], mid, self.cols);
+                slice_2 = MatrixSlice::from_matrix(self, [mid,0], self.rows - mid, self.cols);
             },
             Axes::Col => {
                 assert!(mid < self.cols);
 
-                slice_1 = slice::MatrixSlice::from_matrix(self, [0,0], self.rows, mid);
-                slice_2 = slice::MatrixSlice::from_matrix(self, [0,mid], self.rows, self.cols - mid);
+                slice_1 = MatrixSlice::from_matrix(self, [0,0], self.rows, mid);
+                slice_2 = MatrixSlice::from_matrix(self, [0,mid], self.rows, self.cols - mid);
             }
         }
 
@@ -142,26 +171,26 @@ impl<T> Matrix<T> {
     /// let mut a = Matrix::new(3,3, vec![2.0; 9]);
     /// let (b,c) = a.split_at_mut(1, Axes::Col);
     /// ```
-    pub fn split_at_mut(&mut self, mid: usize, axis: Axes) -> (slice::MatrixSliceMut<T>, slice::MatrixSliceMut<T>) {
+    pub fn split_at_mut(&mut self, mid: usize, axis: Axes) -> (MatrixSliceMut<T>, MatrixSliceMut<T>) {
 
         let mat_cols = self.cols;
         let mat_rows = self.rows;
 
-        let slice_1 : slice::MatrixSliceMut<T>;
-        let slice_2 : slice::MatrixSliceMut<T>;
+        let slice_1 : MatrixSliceMut<T>;
+        let slice_2 : MatrixSliceMut<T>;
 
         match axis {
             Axes::Row => {
                 assert!(mid < self.rows);
 
-                slice_1 = slice::MatrixSliceMut::from_matrix(self, [0,0], mid, mat_cols);
-                slice_2 = slice::MatrixSliceMut::from_matrix(self, [mid,0], mat_rows - mid, mat_cols);
+                slice_1 = MatrixSliceMut::from_matrix(self, [0,0], mid, mat_cols);
+                slice_2 = MatrixSliceMut::from_matrix(self, [mid,0], mat_rows - mid, mat_cols);
             },
             Axes::Col => {
                 assert!(mid < self.cols);
 
-                slice_1 = slice::MatrixSliceMut::from_matrix(self, [0,0], mat_rows, mid);
-                slice_2 = slice::MatrixSliceMut::from_matrix(self, [0,mid], mat_rows, mat_cols - mid);
+                slice_1 = MatrixSliceMut::from_matrix(self, [0,0], mat_rows, mid);
+                slice_2 = MatrixSliceMut::from_matrix(self, [0,mid], mat_rows, mat_cols - mid);
             }
         }
 
@@ -1044,446 +1073,6 @@ impl<T> Matrix<T> where T: Copy + One + Zero + Neg<Output=T> +
         let sgn = p.parity();
 
         sgn * d
-    }
-}
-
-/// Multiplies matrix by scalar.
-impl<T: Copy + One + Zero + Mul<T, Output = T>> Mul<T> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn mul(self, f: T) -> Matrix<T> {
-        self * &f
-    }
-}
-
-/// Multiplies matrix by scalar.
-impl<'a, T: Copy + One + Zero + Mul<T, Output = T>> Mul<&'a T> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn mul(mut self, f: &T) -> Matrix<T> {
-        for val in self.data.iter_mut() {
-            *val = *val * *f;
-        }
-
-        self
-    }
-}
-
-/// Multiplies matrix by scalar.
-impl<'a, T: Copy + One + Zero + Mul<T, Output = T>> Mul<T> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn mul(self, f: T) -> Matrix<T> {
-        self * (&f)
-    }
-}
-
-/// Multiplies matrix by scalar.
-impl<'a, 'b, T: Copy + One + Zero + Mul<T, Output = T>> Mul<&'b T> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn mul(self, f: &T) -> Matrix<T> {
-        let new_data: Vec<T> = self.data.iter().map(|v| (*v) * (*f)).collect();
-
-        Matrix {
-            cols: self.cols,
-            rows: self.rows,
-            data: new_data,
-        }
-    }
-}
-
-/// Multiplies matrix by matrix.
-impl<T: Copy + Zero + One + Mul<T, Output = T> + Add<T, Output = T>> Mul<Matrix<T>> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn mul(self, m: Matrix<T>) -> Matrix<T> {
-        (&self) * (&m)
-    }
-}
-
-/// Multiplies matrix by matrix.
-impl <'a, T: Copy + Zero + One + Mul<T, Output=T> + Add<T, Output=T>> Mul<Matrix<T>> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn mul(self, m: Matrix<T>) -> Matrix<T> {
-        self * (&m)
-    }
-}
-
-/// Multiplies matrix by matrix.
-impl <'a, T: Copy + Zero + One + Mul<T, Output=T> + Add<T, Output=T>> Mul<&'a Matrix<T>> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn mul(self, m: &Matrix<T>) -> Matrix<T> {
-        (&self) * m
-    }
-}
-
-/// Multiplies matrix by matrix.
-impl<'a, 'b, T: Copy + Zero + One + Mul<T, Output=T> + Add<T, Output=T>> Mul<&'b Matrix<T>> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn mul(self, m: &Matrix<T>) -> Matrix<T> {
-        assert!(self.cols == m.rows, "Matrix dimensions do not agree.");
-
-        let mut new_data = vec![T::zero(); self.rows * m.cols];
-
-        unsafe {
-            for i in 0..self.rows
-            {
-                for k in 0..m.rows
-                {
-                    for j in 0..m.cols
-                    {
-                        new_data[i*m.cols() + j] = *new_data.get_unchecked(i*m.cols() + j) + *self.data.get_unchecked(i * self.cols + k) * *m.data.get_unchecked(k*m.cols + j);
-                    }
-                }
-            }
-        }
-
-        Matrix {
-            rows: self.rows,
-            cols: m.cols,
-            data: new_data
-        }
-    }
-}
-
-/// Multiplies matrix by vector.
-impl<T: Copy + Zero + One + Mul<T, Output = T> + Add<T, Output = T>> Mul<Vector<T>> for Matrix<T> {
-    type Output = Vector<T>;
-
-    fn mul(self, m: Vector<T>) -> Vector<T> {
-        (&self) * (&m)
-    }
-}
-
-/// Multiplies matrix by vector.
-impl <'a, T: Copy + Zero + One + Mul<T, Output=T> + Add<T, Output=T>> Mul<Vector<T>> for &'a Matrix<T> {
-    type Output = Vector<T>;
-
-    fn mul(self, m: Vector<T>) -> Vector<T> {
-        self * (&m)
-    }
-}
-
-/// Multiplies matrix by vector.
-impl <'a, T: Copy + Zero + One + Mul<T, Output=T> + Add<T, Output=T>> Mul<&'a Vector<T>> for Matrix<T> {
-    type Output = Vector<T>;
-
-    fn mul(self, m: &Vector<T>) -> Vector<T> {
-        (&self) * m
-    }
-}
-
-/// Multiplies matrix by vector.
-impl<'a, 'b, T: Copy + One + Zero + Mul<T, Output=T> + Add<T, Output=T>> Mul<&'b Vector<T>> for &'a Matrix<T> {
-    type Output = Vector<T>;
-
-    fn mul(self, v: &Vector<T>) -> Vector<T> {
-        assert!(v.size() == self.cols, "Matrix and Vector dimensions do not agree.");
-
-        let mut new_data = Vec::with_capacity(self.rows);
-
-        for i in 0..self.rows
-        {
-            new_data.push(utils::dot(&self.data[i*self.cols..(i+1)*self.cols], v.data()));
-        }
-
-        Vector::new(new_data)
-    }
-}
-
-/// Adds scalar to matrix.
-impl<T: Copy + One + Zero + Add<T, Output = T>> Add<T> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn add(self, f: T) -> Matrix<T> {
-        self + &f
-    }
-}
-
-/// Adds scalar to matrix.
-impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<T> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn add(self, f: T) -> Matrix<T> {
-        self + (&f)
-    }
-}
-
-/// Adds scalar to matrix.
-impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<&'a T> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn add(mut self, f: &T) -> Matrix<T> {
-        for val in self.data.iter_mut() {
-            *val = *val + *f;
-        }
-
-        self
-    }
-}
-
-impl<'a, 'b, T: Copy + One + Zero + Add<T, Output = T>> Add<&'b T> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn add(self, f: &T) -> Matrix<T> {
-        let new_data: Vec<T> = self.data.iter().map(|v| (*v) + (*f)).collect();
-
-        Matrix {
-            cols: self.cols,
-            rows: self.rows,
-            data: new_data,
-        }
-    }
-}
-
-/// Adds matrix to matrix.
-impl<T: Copy + One + Zero + Add<T, Output = T>> Add<Matrix<T>> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn add(self, f: Matrix<T>) -> Matrix<T> {
-        self + &f
-    }
-}
-
-/// Adds matrix to matrix.
-impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<Matrix<T>> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn add(self, f: Matrix<T>) -> Matrix<T> {
-        f + self
-    }
-}
-
-/// Adds matrix to matrix.
-impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<&'a Matrix<T>> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn add(mut self, f: &Matrix<T>) -> Matrix<T> {
-        utils::in_place_vec_bin_op(&mut self.data, &f.data, |x,&y| {*x = *x + y});
-        self
-    }
-}
-
-/// Adds matrix to matrix.
-impl<'a, 'b, T: Copy + One + Zero + Add<T, Output = T>> Add<&'b Matrix<T>> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn add(self, m: &Matrix<T>) -> Matrix<T> {
-        assert!(self.cols == m.cols, "Column dimensions do not agree.");
-        assert!(self.rows == m.rows, "Row dimensions do not agree.");
-
-        let new_data = utils::vec_sum(&self.data, &m.data);
-
-        Matrix {
-            cols: self.cols,
-            rows: self.rows,
-            data: new_data,
-        }
-    }
-}
-
-/// Subtracts scalar from matrix.
-impl<T: Copy + One + Zero + Sub<T, Output = T>> Sub<T> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn sub(self, f: T) -> Matrix<T> {
-        self - (&f)
-    }
-}
-
-/// Subtracts scalar from matrix.
-impl<'a, T: Copy + One + Zero + Sub<T, Output = T>> Sub<&'a T> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn sub(mut self, f: &T) -> Matrix<T> {
-        for val in self.data.iter_mut() {
-            *val = *val - *f;
-        }
-
-        self
-    }
-}
-
-/// Subtracts scalar from matrix.
-impl<'a, T: Copy + One + Zero + Sub<T, Output = T>> Sub<T> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn sub(self, f: T) -> Matrix<T> {
-        self - (&f)
-    }
-}
-
-/// Subtracts scalar from matrix.
-impl<'a, 'b, T: Copy + One + Zero + Sub<T, Output = T>> Sub<&'b T> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn sub(self, f: &T) -> Matrix<T> {
-        let new_data = self.data.iter().map(|v| *v - *f).collect();
-
-        Matrix {
-            cols: self.cols,
-            rows: self.rows,
-            data: new_data,
-        }
-    }
-}
-
-/// Subtracts matrix from matrix.
-impl<T: Copy + One + Zero + Sub<T, Output = T>> Sub<Matrix<T>> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn sub(self, m: Matrix<T>) -> Matrix<T> {
-        self - &m
-    }
-}
-
-/// Subtracts matrix from matrix.
-impl<'a, T: Copy + One + Zero + Sub<T, Output = T>> Sub<Matrix<T>> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn sub(self, mut m: Matrix<T>) -> Matrix<T> {
-        utils::in_place_vec_bin_op(&mut m.data, &self.data, |x,&y| {*x = y - *x});
-        
-        m
-    }
-}
-
-/// Subtracts matrix from matrix.
-impl<'a, T: Copy + One + Zero + Sub<T, Output = T>> Sub<&'a Matrix<T>> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn sub(mut self, m: &Matrix<T>) -> Matrix<T> {
-        utils::in_place_vec_bin_op(&mut self.data, &m.data, |x,&y| {*x = *x - y});
-        
-        self
-    }
-}
-
-/// Subtracts matrix from matrix.
-impl<'a, 'b, T: Copy + One + Zero + Sub<T, Output = T>> Sub<&'b Matrix<T>> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn sub(self, m: &Matrix<T>) -> Matrix<T> {
-        assert!(self.cols == m.cols, "Column dimensions do not agree.");
-        assert!(self.rows == m.rows, "Row dimensions do not agree.");
-
-        let new_data = utils::vec_sub(&self.data, &m.data);
-
-        Matrix {
-            cols: self.cols,
-            rows: self.rows,
-            data: new_data,
-        }
-    }
-}
-
-/// Divides matrix by scalar.
-impl<T: Copy + One + Zero + PartialEq + Div<T, Output = T>> Div<T> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn div(self, f: T) -> Matrix<T> {
-        self / (&f)
-    }
-}
-
-/// Divides matrix by scalar.
-impl<'a, T: Copy + One + Zero + PartialEq + Div<T, Output = T>> Div<T> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn div(self, f: T) -> Matrix<T> {
-        self / (&f)
-    }
-}
-
-/// Divides matrix by scalar.
-impl<'a, T: Copy + One + Zero + PartialEq + Div<T, Output = T>> Div<&'a T> for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn div(mut self, f: &T) -> Matrix<T> {
-        for val in self.data.iter_mut() {
-            *val = *val / *f;
-        }
-
-        self
-    }
-}
-
-/// Divides matrix by scalar.
-impl<'a, 'b, T: Copy + One + Zero + PartialEq + Div<T, Output = T>> Div<&'b T> for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn div(self, f: &T) -> Matrix<T> {
-        assert!(*f != T::zero());
-
-        let new_data = self.data.iter().map(|v| *v / *f).collect();
-
-        Matrix {
-            cols: self.cols,
-            rows: self.rows,
-            data: new_data,
-        }
-    }
-}
-
-/// Gets negative of matrix.
-impl<T: Neg<Output = T> + Copy> Neg for Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn neg(mut self) -> Matrix<T> {
-        for val in self.data.iter_mut() {
-            *val = -*val;
-        }
-
-        self
-    }
-}
-
-/// Gets negative of matrix.
-impl<'a, T: Neg<Output = T> + Copy> Neg for &'a Matrix<T> {
-    type Output = Matrix<T>;
-
-    fn neg(self) -> Matrix<T> {
-        let new_data = self.data.iter().map(|v| -*v).collect();
-
-        Matrix {
-            cols: self.cols,
-            rows: self.rows,
-            data: new_data,
-        }
-    }
-}
-
-/// Indexes matrix.
-///
-/// Takes row index first then column.
-impl<T> Index<[usize; 2]> for Matrix<T> {
-    type Output = T;
-
-    fn index(&self, idx: [usize; 2]) -> &T {
-        assert!(idx[0] < self.rows,
-                "Row index is greater than row dimension.");
-        assert!(idx[1] < self.cols,
-                "Column index is greater than column dimension.");
-        unsafe { &self.data.get_unchecked(idx[0] * self.cols + idx[1]) }
-    }
-}
-
-/// Indexes mutable matrix.
-///
-/// Takes row index first then column.
-impl<T> IndexMut<[usize; 2]> for Matrix<T> {
-
-    fn index_mut(&mut self, idx: [usize; 2]) -> &mut T {
-        assert!(idx[0] < self.rows,
-                "Row index is greater than row dimension.");
-        assert!(idx[1] < self.cols,
-                "Column index is greater than column dimension.");
-        let self_cols = self.cols;
-        unsafe { self.data.get_unchecked_mut(idx[0] * self_cols + idx[1]) }
     }
 }
